Move D2GSWatchDogInit(), D2GSResetWatchDogCounter(), CheckWatchDogCounter(), and D2GSWatchDogThread() from main.cpp to watchdog.cpp
