+++ src/d2gs/watchdog.h
@@ -1,0 +1,7 @@
+#ifndef INCLUDED_WATCHDOG_H
+#define INCLUDED_WATCHDOG_H
+
+void D2GSWatchDogInit(void);
+void D2GSResetWatchDogCounter(void);
+
+#endif
